+++ src/app/admin/[eventId]/components/VotingDashboard/index.tsx
