+++ src/app/admin/[eventId]/hooks/useEventAdmin.ts
