+++ src/app/admin/components/DeleteEvent.tsx
