+++ src/app/admin/present/contexts/PresentationContext.ts
@@ -2,7 +2,7 @@ import { createContext, useContext } from "react";
 
 import { type CurrentEvent } from "~/lib/types/currentEvent";
 
-import { type AdminEvent } from "~/app/admin/contexts/DashboardContext";
+import { type AdminEvent } from "~/app/admin/[eventId]/contexts/DashboardContext";
 
 export type IPresentationContext = {
   currentEvent: CurrentEvent;
